Count references/citations from files

count_ref() count references/citations from files. If more than one file is used, the function returns the sum of the count of the references/citations from each file.

At the moment, count_ref() works only with PubMed and RIS (Research Information Systems) formats.

count_ref(file = file.choose())

Arguments

file
(optional) a character object indicating RIS or ZIP file names. If not assigned, a dialog window will be open enabling the user to search and select a file (only for interactive sessions).

Value

A numeric value with the count or sum of the count of the references/citations found in the file argument.

Details

count_ref() also allows you to read ZIP compacted files. Just assign the ZIP file in the file argument.

See also

Examples

if (FALSE) {
citations <- raw_data("citation")
file <- citations[grep("_apa_", citations)]
file <- raw_data("citation", file)

count_ref(file)}