Group keywords from the keyword dataset

CAUTION: This function must be used only with packages that follow the SQLR system.

keyword_set() returns a keyword set from the keyword dataset.

keyword_set(domain_id, language = NULL, package = gutils:::get_package_name())

Arguments

domain_id

An integer or numeric value indicating the domain ID to return.

language

(optional) A string indicating the language constraint of the keywords (case insensitive).

package

(optional) a string indicating the package with the database data. If not assigned, the function will try to use the name of the active project directory (requires the rstudioapi package).

Value

A character object with a keyword set.

Examples

if (FALSE) {
keyword_set(1, "english")}