Print statistics about the selection process

CAUTION: This function must be used only with packages that follow the SQLR system.

selection_stats() prints statistics about a trial or the selection process as a whole.

The trial statistics are computed by downloading the trial table (requires an internet connection). The full statistics are computed using the reference data of a SQLR package, hence, the data of the package must be up-to-date.

selection_stats(
  package = gutils:::get_package_name(),
  trial_id = NULL,
  clipboard = TRUE
)

Arguments

package

(optional) a string indicating the package with the database data. If not assigned, the function will try to use the name of the active project directory (requires the rstudioapi package).

trial_id

(optional) a string indicating the ID of the trial. This argument must be set only when the function must print a trial statistics (default: NULL).

clipboard

(optional) a logical value indicating if the function must copy a Markdown version of the statistics to the clipboard. (default: TRUE).

Examples

if (FALSE) {
selection_stats()}