Start settings for SQLR R package templates

start_sqlr() handles the initial commands to start systematic quantitative literature reviews R packages.

start_sqlr(id, sheet = "Dataset", package = rutils:::get_package_name())

Arguments

id

A string with the Google Sheets ID from the 'Sheets' table.

sheet

(optional) a string indicating the worksheet/tab where the sheets data can be found on the 'Sheets spreadsheet (default: "Dataset").

package

(optional) a string indicating the package with the database data. If not assigned, the function will try to use the name of the active project directory (requires the rstudioapi package).