Create a list with Google Sheets metadata for tables hosted on the platform

CAUTION: This function must be used only with packages that follow the SQLR system.

write_metadata() creates and returns a list object containing lists with the Google Sheets metadata of the review tables hosted on the platform.

This must be set for all R packages from GIPSO for systematic quantitative literature reviews.

You can see a example of a metadata sheets in: https://bit.ly/2PFWhev.

write_metadata(id, sheet = "Dataset")

Arguments

id

A string with the Google Sheets ID from the Sheets table.

sheet

(optional) a string indicating the worksheet/tab where the sheets data can be found in the Sheets spreadsheet (default: "Dataset").

Value

An invisible list object containing lists with the Google Sheets metadata of the review tables hosted on the platform.

See also

Examples

if (FALSE) {
write_metadata("1x6Aj8cXl9qFtpXq48Q6zdmT-w9EEqNG1iVQcKVRhgKM")
}