Build and write the reference table to a SQLR package

build_reference() builds and write the reference entity table of the Systematic Quantitative Literature Review (SQLR) system to an R package.

You must have a sheets data object with the sheets metadata before running this function. See write_metadata() to learn more.

build_reference(package = gutils:::get_package_name(), write = TRUE)

Arguments

package
(optional) a string indicating the package with the database data. If not assigned, the function will try to use the name of the active project directory (requires the rstudioapi package).

write
(optional) a logical value indicating if the function must write a reference.rda file to "./data/" and also write to the reference spreadsheet listed on the sqlr::sheets object (default: TRUE).

See also

Examples

if (FALSE) {
build_reference()}