Group keywords from the keyword dataset

keyword_set() returns a keyword set from the keyword dataset.

keyword_set(domain_id, language = NULL, package = NULL)

Arguments

domain_id
An integer or numeric value indicating the domain ID to return.

language
(optional) A string indicating the language constraint of the keywords (case insensitive).

package
(optional) a string indicating the package with the database data. If NULL, the function will try to use the name of the active project directory (requires the rstudioapi package) (default: NULL).

Value

A character object with a keyword set.

See also

Examples

if (FALSE) {
keyword_set(1, "english")}